+++ packages/kit/src/runtime/server/cookie.spec.js
@@ -1,6 +1,6 @@
 import { test } from 'uvu';
 import * as assert from 'uvu/assert';
-import { domain_matches, path_matches } from './cookie.js';
+import { domain_matches, path_matches, get_cookies } from './cookie.js';
 
 const domains = {
 	positive: [
@@ -37,6 +37,19 @@ paths.negative.forEach(([path, constraint]) => {
 	test(`! ${path} / ${constraint}`, () => {
 		assert.ok(!path_matches(path, constraint));
 	});
+});
+
+test('a cookie should not be present after it is deleted', () => {
+	const url = new URL('https://example.com');
+	const request = new Request(new URL('https://example.com'), {
+		headers: new Headers({
+			cookie: 'a=b;'
+		})
+	});
+	const { cookies } = get_cookies(request, url);
+	assert.equal(cookies.get('a'), 'b');
+	cookies.delete('a');
+	assert.not(cookies.get('a'));
 });
 
 test.run();